Fixes variable hiding in bootstrapper
Even though it may be better to have the board passed as an argument to
drawing functions, as it currently stands it just hid the board var in
the same scope. Better to remove this source of confusion and bring back
the function argument later when redrawing would be moved elsewhere
anyway.