Introduces rudimentary undo/redo/rematch controls
The reason for going with this rather crude version of binding the view
to application logic is simply "First make it work, then make it work
better". I'm not happy because the code's repetitive, the bootstrapper
is getting too many responsibilities, and in general not very DRY.
However, it's better to have something than nothing.

This commit also makes a Board a true dependency of Match, where
starting a fresh match will clear the board.