Introduces "winning" in the view and bootstrapper
More rudimentary bootstrapping necessary to get a first full working
version of the application.

This commit does make one problem clear: the distinction between a
board and a match is not as clear in code as it could be. Particularly
troublesome is the fact that a Match has methods for finding winners,
which make it partially a facade for the board (which has the actual
logic about winners). This should be considered an open issue.