Introduces poor-man's pub/sub for state changes
For now we're trying to keep this project (a) free from external libs as
much as possible, and (b) limited in the amount of invested time. This
is why we introduce a poor-man's version of pub/sub (or event handling)
to handle state changes. In the end this should make the data binding
code much clearer / DRYer.