Implements the default DAL with localStorage
This is a front-end test project (for now), so we use a DAL that uses
localStorage as a default. The DAL is a dependency of MatchRepository
objects, and a version that uses Ajax to communicate with a server
should be easy to inject.

Note that this commit still has a gruesome bug with saving matches, the
replay will *not* yet be correct. Fixes are forthcoming.