Rewrites MatchRepository to AngularJS service
The repository is better suited as an AngularJS service than as a domain
entity. This also requires all functionality dependent on repositories
to change.

In addition, since this is an experimentation repo with no MAKE files or
other type of build scripts, we may as well include the unminified
AngularJS lib to get better feedback on injection problems.

Finally, this also moves the app startup functionality to a separate
piece, so that it's not included when running tests.

For more info on how the unit tests for the repository service are
written, see http://codereview.stackexchange.com/q/98519/10034